Rolled back the change to the scanner from the previous commit, and made a breaking change to the bundle() function in o.expr.
The change to the scanner ended up creating all kinds of parsing
problems, so for now, the old behavior has been restored.

Prior to this commit, the bundle() function was a special function--
its arguments were not evaluated, and then it expected to find OSC
addresses in the odd argument positions, and would proceed to evaluate
the even-position arguments. It is now a regular function, which means
that all arguments are evaluated, so addresses must be passed as strings
(in quotes or using the quote() function). This was done so that
bundle() could handle a single argument case of either a list or a blob,
that would be converted into a subbundle. The motivation for this was
integration with Vanilla PD--the netsend/receive objects take and produce
lists of bytes, and now bundle() will convert a list of bytes into
a bundle.