Rolled back an over-aggressive parameter check for 2-arg functions in the expr language.
Prior to 3b64c3c67a, a 2-arg function could be called with an empty
second argument, like this:

/x = [],
/y = /x + 3

which would produce /y bound to 3. I recognized this as a bug, but
after discussions with Jon Kulpa, who was using this behavior to
do a kind of initialization, I now think of this as a feature.

The interpretation of the expression above is "don't add anything
to 3".

In any case, it's easy to guard against this behavior if necessary
using bound() and exists().